Fixed variable datatype.

+++ libosmscout/src/osmscout/Way.cpp
@@ -203,7 +203,7 @@ namespace osmscout {
 
       while (idCurrent<ids.size()) {
         uint8_t bitset=0;
-        size_t  bitMask=1;
+        uint8_t bitMask=1;
         size_t  idEnd=std::min(idCurrent+8,ids.size());
 
         for (size_t i=idCurrent; i<idEnd; i++) {
